more_timing_updates
